♻️ notice api요청 리팩토링
